Update 08_Lucky unicorn.py
What you have is great.  I have made a few changes to the token_statement function so that it is possible to include the feedback and the users' balance inside the decoration.  This also meant editing the main routine a bit.

Please compare what you have with this trial / iteration.  Perhaps it will inspire you when you create your next program.